Add a test class for trait UsePusherChannelsNames
